requirements.txt
psutil
speedtest-cli
py-cpuinfo
pywin32
winshell
urllib3
